Make it possible to specify custom hipchat server
As hipchat can be installed on own server, we need an option to call
the API that is on a custom server. So, the end point needs to support
specifying custom hostname. Specify the host name in
hipchat_endpoint_host. If not present, default of api.hipchat.com will
be used.